Disable custom LESS arguments due to conflict
Setting the custom arguments as I did seems to conflict with 'git diff'.
Expected output is a colorized diff of the content differences. Actual
output is a diff with color codes printed directly into the terminal but
not applied to the text.

Reverting the change for now, need to research this issue and determine
if other arguments exist to correct this. Or if there is an alternative
method to allowing mouse-wheel in 'less' that will not conflict with the
color codes.