Add `burnme` alias
This alias will `grep` the output of `lm-sensors` for lines beginning
with core in a case-insensitive fashion. The rationale behind this alias
was out of the 36 lines of output from `sensors -f` on my machine, I am
only looking at 2-4 of them. This alias will save scrollback buffer, a
highly underrated commoditiy.

It also has a fun name designed for minimal keystrokes.