`ehowlong` alias - How long did it take?
The `ehowlong` alias accepts a package name, or names, and outputs a
listing of the date, time, version and build times for each time that
package was installed through Portage.

Relatively speaking, this command saves no keystrokes. It mostly allows
me to tab complete after two keystrokes, and is a bit easier for me to
remember than `sudo genlop -t`.