Add global .gitignore
Finally decided it would be worth the time invested to set up a global
`.gitignore` file for our user account. This will eliminate my need to
write a `.git/info/exclude` file on each repo I create. This file only
really blocks a couple files, but they exist in every repo I work on.
Makes sense to block on a per-user instead of per-repo basis.

Due to the fact that these files are contained in a repo, with their own
`.gitignore` file, we needed a different filename. I chose
`.gitignore.global` as it maintains *what* the file is, and indicates
it's status as meant for global use.

The 'installation' of this file is much the same as other files in this
repo. One key difference though -- you must tell `git` about this new
file, and where it is located. Commands such as the following should be
enough to get it properly placed and working:

```
git config --global core.excludesfile '~/.gitignore'
ln -s /path/to/repo/.gitignore.global /home/username/.gitignore
```