Remove ConsoleKit support in favor of elogind.
As per the Gentoo migration away from ConsoleKit support in the 17.1
Desktop profile, I am updating this script to no longer depend on
ConsoleKit/`ck-launch-session`. That executable is slated for removal
along with ConsoleKit as I proceed into this update.

Under normal circumstances, this `.xinitrc` script will *not* be called.
My standard login is through `sddm` which loads sessions from
`/etc/X11/Sessions`, thus the `awesome` script inside that location is
what will actually be called. This update is simply to prevent this
script from being broken in the future if I ever need to `startx`
manually. If it saves me at least one headache, or Duck Search, it was
worth it.