Include autorun script for AwesomeWM
This script handles some startup programs for AwesomeWM which will be
launched at session start. Some of this is to set up the environment for
use by Awesome (font path, color profile, etc). Others are used for
launching certain programs I wish to load automatically at each new
session.

The script is designed such that the auto-start programs which stay
running in the background, or as tray icons, will only be executed if
there is not already a running instance.

The reference for this implementation was taken from the ArchWiki:
https://wiki.archlinux.org/index.php/Awesome#Autostart