Migrate to new auto-start setup
Migrate the auto-start system from the deprecated `awful.util.spawn`
method in favor of the `awful.spawn.with_shell` method.

The latter worked initially, but after an update certain programs would
not execute when the session started. This resulted in the inability to
launch a `urxvt` terminal, as it was unable to find the proper font. I
was also unhappy with keeping a list of shell commands inside of the
`rc.lua` file but was unsure of another method that actually worked for
my setup.

Thus, we have migrated the code in this `rc.lua` file to call on a newly
added `autorun.sh` script. Testing shows this works, but I'll keep an
eye on it for now.

This commit resolves issue #1.