Enable parinfer, dired icons
Parinfer mode has been handling *very* nicely as a mostly automated
means to prevent insanity from parentheses in Lisp editing modes. Very
nice implementation, and so far pain free.

Enable `all-the-icons` for Dired mode. Just a little eye-candy. Not sure
I want to keep it, but it is helpful at least for now.