Update monitor refresh rate to 120Hz
Noticed a touch of tearing here and there while playing games with the
fps locked to 60. Due to the previous monitor refresh rate being 100Hz,
that may cause tearing as the video output and monitor refresh will be
out of sync. Bumping the refresh rate up to 120Hz should handle those
bits of tearing.