Change desktop wallpaper
Switch back over to an older wallpaper, just changing it up a bit.

I really should tweak this bit of config so every time I change the
wallpaper, a commit is not required. Possibly use a random selection of
wallpaper on each login. Maybe a cronjob to change the wallpaper if I'm
logged in. Could use a `.gitignore`'d file with the wallpaper value and
read from that.

Lots of options!