Prettify temperature conversion output
Felt it might look nice to include a degrees glyph and temperature unit
into the output of the temperature conversions introduced in the
previous commit. I like it.

The little Python one-liner is getting a little harder to follow and
doesn't look the nicest. Might play around with it more later. Try to
get it looking a little neater. We'll see.