Weather widget added to bottom panel
Hooked up a little code to call my `weather_cli` script to output the
current weather conditions onto the bottom panel. Weather is updated
every 30 minutes. The command is a simple one-liner, however I am
running it through an external script. This is basically so I can not
commit my location into the git repo :)

Otherwise, for reference, the command is:

  $ weather current -u imperial -Cthw -p short [LOCATION]