Include dunst config file
Figured I'm trying a new window manager, may as well try a new
notification server as well -- Dunst! Took a little messing around to
get set up, and to replace notification-daemon but it seems to work
perfectly fine now.

The config is based upon the default included with dunst at
`/etc/xdg/dunst/dunstrc`. I did not think to make a commit of the
unmodified default, so this commit does include my initial changes to
the config. Really not much, so I'm not very concerned about it.

Further steps will be themeing and tweaking as necessary. Liking dunst
so far though.