Reconfigure panel bar widgets
Moved things around into what I think is a bit more organized of a setup
overall. The new 'flow' will keep widgets and information most commonly
used on the top panel. And the bottom panel will be more for 'status'
type messages. I think this will help keep my eyes from jumping back and
forth from top to bottom. More quickly able to hone in on the
information I want.

To put more simply: up top we have the group listing, focused window
title, systray, and clock. On the bottom we have the current layout name,
CPU core temperatures, potential updates counter, and weather.

As part of this, we also prefixed group labels with a number. This is
purely to help me with hotkeys and knowing which group is which number
on the keyboard.