Auto-set wallpaper on login
Kept forgetting to have the wallpaper set at login. Due to the way
`autorun.sh` is set up, with wallpaper set from within AwesomeWM, I did
not want to override that and set the Qtile wallpaper within the script.

Thus I have settled on an alternative solution. The wallpaper is defined
in a variable at the top of Qtile config file, and then `feh` is called
upon to set the wallaper via a hook into the `startup_once` method.

I haven't tested this directly. But Python doesn't complain when I
compile to bytecode. Just don't particularly feel like re-logging in
fully right now. Yet in the time to write this paragraph I could have
done so. Either way, should work. We'll know for sure eventually!