Fix update counter widget
I have worked out a solution to the 'potential updates' widget having an
incorrect number. This solution comes at the cost of increased cpu use
though, so there is a tradeoff. And I may adjust the update interval to
be longer in the future for that reason.

Rather than calling upon `eix`, we run `emerge` in pretend mode to avoid
needing elevated permissions. Suppress the spinner, and ask for quiet
output. We get one package name per line if Portage would like to merge
it for whichever reason. This is the true number I am looking for. Just
costs a little more power to get it.