add new alias `einfo`
This alias simply calls upon `genlop -i` to provide a nicely detailed
information blob about an ebuild/package. Halfway thought to use this to
replace the `ehowlong` alias, but decided to go for a separate solution
for now. Each has a specific purpose, but I think I would get more use
from the `einfo` alias set here. This is a really long commit message
for just an alias. Thanks for reading, have a nice day.